Moved the WonkaEth.Orchestration events and outputs into subfolders.
