+++ WonkaSystem/WonkaEth/Orchestration/WonkaOrchestratorException.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 
+using WonkaEth.Orchestration.BlockchainOutput;
+
 namespace WonkaEth.Orchestration
 {
     public class WonkaOrchestratorException : Exception
